chore: tease out yggdrasil adapters into their own static class
